--- src/interfaces/assistants_web/.storybook/preview.ts
@@ -1,22 +1,0 @@
-import type { Preview } from '@storybook/react';
-
-import '../src/styles/main.css';
-
-const preview: Preview = {
-  parameters: {
-    nextjs: {
-      appDirectory: true,
-      router: {
-        basePath: '/',
-      },
-    },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
