+++ src/interfaces/assistants_web/src/components/Shared/Banner/Banner.stories.tsx
@@ -1,7 +1,7 @@
 // Replace your-framework with the name of your framework
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Banner } from './Banner';
+import { Banner } from '@/components/Shared';
 
 const meta: Meta<typeof Banner> = {
   title: 'Components/Banner',
